Openclaw_RAG/requirements.txt
langchain
langchain-openai
chromadb
python-dotenv
tiktoken
langchain-chroma
